requirements.txt
discord.py==2.3.2
firebase-admin==6.5.0
python-dotenv==1.0.1
google-generativeai==0.4.1